+++ src/app/cli/src/cli_commands.rs
@@ -19,46 +19,46 @@ use crate::{accounts, cli, odf_server, WorkspaceService};
 
 pub fn get_command(
     base_catalog: &Catalog,
-    work_catalog: &Catalog,
+    cli_catalog: &Catalog,
     args: cli::Cli,
 ) -> Result<Box<dyn Command>, CLIError> {
     let command: Box<dyn Command> = match args.command {
         cli::Command::Add(c) => Box::new(AddCommand::new(
-            work_catalog.get_one()?,
-            work_catalog.get_one()?,
-            work_catalog.get_one()?,
-            work_catalog.get_one()?,
-            work_catalog.get_one()?,
+            cli_catalog.get_one()?,
+            cli_catalog.get_one()?,
+            cli_catalog.get_one()?,
+            cli_catalog.get_one()?,
+            cli_catalog.get_one()?,
             c.manifest,
             c.name,
             c.recursive,
             c.replace,
             c.stdin,
             c.visibility.into(),
-            work_catalog.get_one()?,
+            cli_catalog.get_one()?,
         )),
         cli::Command::Complete(c) => {
-            let workspace_svc = work_catalog.get_one::<WorkspaceService>()?;
+            let workspace_svc = cli_catalog.get_one::<WorkspaceService>()?;
             let in_workspace =
                 workspace_svc.is_in_workspace() && !workspace_svc.is_upgrade_needed()?;
 
             Box::new(CompleteCommand::new(
                 if in_workspace {
-                    Some(work_catalog.get_one()?)
+                    Some(cli_catalog.get_one()?)
                 } else {
                     None
                 },
                 if in_workspace {
-                    Some(work_catalog.get_one()?)
+                    Some(cli_catalog.get_one()?)
                 } else {
                     None
                 },
                 if in_workspace {
-                    Some(work_catalog.get_one()?)
+                    Some(cli_catalog.get_one()?)
                 } else {
                     None
                 },
-                work_catalog.get_one()?,
+                cli_catalog.get_one()?,
                 cli::Cli::command(),
                 c.input,
                 c.current,
@@ -69,39 +69,39 @@ pub fn get_command(
         }
         cli::Command::Config(c) => match c.subcommand {
             cli::ConfigSubCommand::List(sc) => Box::new(ConfigListCommand::new(
-                work_catalog.get_one()?,
+                cli_catalog.get_one()?,
                 sc.user,
                 sc.with_defaults,
             )),
             cli::ConfigSubCommand::Get(sc) => Box::new(ConfigGetCommand::new(
-                work_catalog.get_one()?,
+                cli_catalog.get_one()?,
                 sc.user,
                 sc.with_defaults,
                 sc.cfgkey,
             )),
             cli::ConfigSubCommand::Set(sc) => Box::new(ConfigSetCommand::new(
-                work_catalog.get_one()?,
+                cli_catalog.get_one()?,
                 sc.user,
                 sc.cfgkey,
                 sc.value,
             )),
         },
         cli::Command::Delete(c) => Box::new(DeleteCommand::new(
-            work_catalog.get_one()?,
-            work_catalog.get_one()?,
-            work_catalog.get_one()?,
-            validate_many_dataset_patterns(work_catalog, c.dataset)?,
-            work_catalog.get_one()?,
+            cli_catalog.get_one()?,
+            cli_catalog.get_one()?,
+            cli_catalog.get_one()?,
+            validate_many_dataset_patterns(cli_catalog, c.dataset)?,
+            cli_catalog.get_one()?,
             c.all,
             c.recursive,
         )),
         cli::Command::Ingest(c) => Box::new(IngestCommand::new(
-            work_catalog.get_one()?,
-            work_catalog.get_one()?,
-            work_catalog.get_one()?,
-            work_catalog.get_one()?,
-            work_catalog.get_one()?,
-            validate_dataset_ref(work_catalog, c.dataset)?,
+            cli_catalog.get_one()?,
+            cli_catalog.get_one()?,
+            cli_catalog.get_one()?,
+            cli_catalog.get_one()?,
+            cli_catalog.get_one()?,
+            validate_dataset_ref(cli_catalog, c.dataset)?,
             c.file.unwrap_or_default(),
             c.source_name,
             c.event_time,
@@ -112,15 +112,15 @@ pub fn get_command(
         cli::Command::Init(c) => {
             if c.pull_images {
                 Box::new(PullImagesCommand::new(
-                    work_catalog.get_one()?,
-                    work_catalog.get_one()?,
-                    work_catalog.get_one()?,
+                    cli_catalog.get_one()?,
+                    cli_catalog.get_one()?,
+                    cli_catalog.get_one()?,
                     c.list_only,
                 ))
             } else {
                 Box::new(InitCommand::new(
-                    work_catalog.get_one()?,
-                    work_catalog.get_one()?,
+                    cli_catalog.get_one()?,
+                    cli_catalog.get_one()?,
                     c.exists_ok,
                     c.multi_tenant,
                 ))
@@ -128,34 +128,34 @@ pub fn get_command(
         }
         cli::Command::Inspect(c) => match c.subcommand {
             cli::InspectSubCommand::Lineage(sc) => Box::new(LineageCommand::new(
-                work_catalog.get_one()?,
-                work_catalog.get_one()?,
-                work_catalog.get_one()?,
-                validate_many_dataset_refs(work_catalog, sc.dataset)?,
+                cli_catalog.get_one()?,
+                cli_catalog.get_one()?,
+                cli_catalog.get_one()?,
+                validate_many_dataset_refs(cli_catalog, sc.dataset)?,
                 sc.browse,
                 sc.output_format,
-                work_catalog.get_one()?,
+                cli_catalog.get_one()?,
             )),
             cli::InspectSubCommand::Query(sc) => Box::new(InspectQueryCommand::new(
-                work_catalog.get_one()?,
-                work_catalog.get_one()?,
-                validate_dataset_ref(work_catalog, sc.dataset)?,
-                work_catalog.get_one()?,
+                cli_catalog.get_one()?,
+                cli_catalog.get_one()?,
+                validate_dataset_ref(cli_catalog, sc.dataset)?,
+                cli_catalog.get_one()?,
             )),
             cli::InspectSubCommand::Schema(sc) => Box::new(InspectSchemaCommand::new(
-                work_catalog.get_one()?,
-                validate_dataset_ref(work_catalog, sc.dataset)?,
+                cli_catalog.get_one()?,
+                validate_dataset_ref(cli_catalog, sc.dataset)?,
                 sc.output_format,
                 sc.from_data_file,
             )),
         },
         cli::Command::List(c) => {
-            let workspace_svc = work_catalog.get_one::<WorkspaceService>()?;
-            let user_config = work_catalog.get_one::<kamu_accounts::PredefinedAccountsConfig>()?;
+            let workspace_svc = cli_catalog.get_one::<WorkspaceService>()?;
+            let user_config = cli_catalog.get_one::<kamu_accounts::PredefinedAccountsConfig>()?;
 
             Box::new(ListCommand::new(
-                work_catalog.get_one()?,
-                work_catalog.get_one()?,
+                cli_catalog.get_one()?,
+                cli_catalog.get_one()?,
                 accounts::AccountService::current_account_indication(
                     args.account,
                     workspace_svc.is_multi_tenant_workspace(),
@@ -165,23 +165,23 @@ pub fn get_command(
                     c.target_account,
                     c.all_accounts,
                 ),
-                work_catalog.get_one()?,
+                cli_catalog.get_one()?,
                 c.wide,
             ))
         }
         cli::Command::Log(c) => Box::new(LogCommand::new(
-            work_catalog.get_one()?,
-            work_catalog.get_one()?,
-            validate_dataset_ref(work_catalog, c.dataset)?,
+            cli_catalog.get_one()?,
+            cli_catalog.get_one()?,
+            validate_dataset_ref(cli_catalog, c.dataset)?,
             c.output_format,
             c.filter,
             c.limit,
-            work_catalog.get_one()?,
+            cli_catalog.get_one()?,
         )),
         cli::Command::Login(c) => match c.subcommand {
             Some(cli::LoginSubCommand::Oauth(sc)) => Box::new(LoginSilentCommand::new(
-                work_catalog.get_one()?,
-                work_catalog.get_one()?,
+                cli_catalog.get_one()?,
+                cli_catalog.get_one()?,
                 if sc.user {
                     odf_server::AccessTokenStoreScope::User
                 } else {
@@ -194,8 +194,8 @@ pub fn get_command(
                 }),
             )),
             Some(cli::LoginSubCommand::Password(sc)) => Box::new(LoginSilentCommand::new(
-                work_catalog.get_one()?,
-                work_catalog.get_one()?,
+                cli_catalog.get_one()?,
+                cli_catalog.get_one()?,
                 if sc.user {
                     odf_server::AccessTokenStoreScope::User
                 } else {
@@ -208,9 +208,9 @@ pub fn get_command(
                 }),
             )),
             None => Box::new(LoginCommand::new(
-                work_catalog.get_one()?,
-                work_catalog.get_one()?,
-                work_catalog.get_one()?,
+                cli_catalog.get_one()?,
+                cli_catalog.get_one()?,
+                cli_catalog.get_one()?,
                 if c.user {
                     odf_server::AccessTokenStoreScope::User
                 } else {
@@ -222,7 +222,7 @@ pub fn get_command(
             )),
         },
         cli::Command::Logout(c) => Box::new(LogoutCommand::new(
-            work_catalog.get_one()?,
+            cli_catalog.get_one()?,
             if c.user {
                 odf_server::AccessTokenStoreScope::User
             } else {
@@ -238,10 +238,10 @@ pub fn get_command(
             None::<&str>,
         )),
         cli::Command::Notebook(c) => Box::new(NotebookCommand::new(
-            work_catalog.get_one()?,
-            work_catalog.get_one()?,
-            work_catalog.get_one()?,
-            work_catalog.get_one()?,
+            cli_catalog.get_one()?,
+            cli_catalog.get_one()?,
+            cli_catalog.get_one()?,
+            cli_catalog.get_one()?,
             c.address,
             c.http_port,
             c.env.unwrap_or_default(),
@@ -249,9 +249,9 @@ pub fn get_command(
         cli::Command::Pull(c) => {
             if let Some(set_watermark) = c.set_watermark {
                 Box::new(SetWatermarkCommand::new(
-                    work_catalog.get_one()?,
-                    work_catalog.get_one()?,
-                    work_catalog.get_one()?,
+                    cli_catalog.get_one()?,
+                    cli_catalog.get_one()?,
+                    cli_catalog.get_one()?,
                     c.dataset.unwrap_or_default(),
                     c.all,
                     c.recursive,
@@ -259,12 +259,12 @@ pub fn get_command(
                 ))
             } else {
                 Box::new(PullCommand::new(
-                    work_catalog.get_one()?,
-                    work_catalog.get_one()?,
-                    work_catalog.get_one()?,
-                    work_catalog.get_one()?,
+                    cli_catalog.get_one()?,
+                    cli_catalog.get_one()?,
+                    cli_catalog.get_one()?,
+                    cli_catalog.get_one()?,
                     c.dataset.unwrap_or_default(),
-                    work_catalog.get_one()?,
+                    cli_catalog.get_one()?,
                     c.all,
                     c.recursive,
                     c.fetch_uncacheable,
@@ -276,52 +276,52 @@ pub fn get_command(
             }
         }
         cli::Command::Push(c) => Box::new(PushCommand::new(
-            work_catalog.get_one()?,
-            work_catalog.get_one()?,
-            work_catalog.get_one()?,
+            cli_catalog.get_one()?,
+            cli_catalog.get_one()?,
+            cli_catalog.get_one()?,
             c.dataset.unwrap_or_default(),
-            work_catalog.get_one()?,
+            cli_catalog.get_one()?,
             c.all,
             c.recursive,
             !c.no_alias,
             c.force,
             c.to,
             c.visibility.into(),
-            work_catalog.get_one()?,
+            cli_catalog.get_one()?,
         )),
         cli::Command::Rename(c) => Box::new(RenameCommand::new(
-            work_catalog.get_one()?,
-            validate_dataset_ref(work_catalog, c.dataset)?,
+            cli_catalog.get_one()?,
+            validate_dataset_ref(cli_catalog, c.dataset)?,
             c.name,
         )),
         cli::Command::Repo(c) => match c.subcommand {
             cli::RepoSubCommand::Add(sc) => Box::new(RepositoryAddCommand::new(
-                work_catalog.get_one()?,
+                cli_catalog.get_one()?,
                 sc.name,
                 sc.url,
             )),
             cli::RepoSubCommand::Delete(sc) => Box::new(RepositoryDeleteCommand::new(
-                work_catalog.get_one()?,
-                work_catalog.get_one()?,
+                cli_catalog.get_one()?,
+                cli_catalog.get_one()?,
                 sc.repository.unwrap_or_default(),
                 sc.all,
             )),
             cli::RepoSubCommand::List(_) => Box::new(RepositoryListCommand::new(
-                work_catalog.get_one()?,
-                work_catalog.get_one()?,
+                cli_catalog.get_one()?,
+                cli_catalog.get_one()?,
             )),
             cli::RepoSubCommand::Alias(sc) => match sc.subcommand {
                 cli::RepoAliasSubCommand::Add(ssc) => Box::new(AliasAddCommand::new(
-                    work_catalog.get_one()?,
-                    work_catalog.get_one()?,
+                    cli_catalog.get_one()?,
+                    cli_catalog.get_one()?,
                     ssc.dataset,
                     ssc.alias,
                     ssc.pull,
                     ssc.push,
                 )),
                 cli::RepoAliasSubCommand::Delete(ssc) => Box::new(AliasDeleteCommand::new(
-                    work_catalog.get_one()?,
-                    work_catalog.get_one()?,
+                    cli_catalog.get_one()?,
+                    cli_catalog.get_one()?,
                     ssc.dataset,
                     ssc.alias,
                     ssc.all,
@@ -329,33 +329,33 @@ pub fn get_command(
                     ssc.push,
                 )),
                 cli::RepoAliasSubCommand::List(ssc) => Box::new(AliasListCommand::new(
-                    work_catalog.get_one()?,
-                    work_catalog.get_one()?,
-                    work_catalog.get_one()?,
+                    cli_catalog.get_one()?,
+                    cli_catalog.get_one()?,
+                    cli_catalog.get_one()?,
                     ssc.dataset,
                 )),
             },
         },
         cli::Command::Reset(c) => Box::new(ResetCommand::new(
-            work_catalog.get_one()?,
-            work_catalog.get_one()?,
-            work_catalog.get_one()?,
-            validate_dataset_ref(work_catalog, c.dataset)?,
+            cli_catalog.get_one()?,
+            cli_catalog.get_one()?,
+            cli_catalog.get_one()?,
+            validate_dataset_ref(cli_catalog, c.dataset)?,
             c.hash,
         )),
         cli::Command::Search(c) => Box::new(SearchCommand::new(
-            work_catalog.get_one()?,
-            work_catalog.get_one()?,
+            cli_catalog.get_one()?,
+            cli_catalog.get_one()?,
             c.query,
             c.repo.unwrap_or_default(),
         )),
         cli::Command::Sql(c) => match c.subcommand {
             None => Box::new(SqlShellCommand::new(
-                work_catalog.get_one()?,
-                work_catalog.get_one()?,
-                work_catalog.get_one()?,
-                work_catalog.get_one()?,
-                work_catalog.get_one()?,
+                cli_catalog.get_one()?,
+                cli_catalog.get_one()?,
+                cli_catalog.get_one()?,
+                cli_catalog.get_one()?,
+                cli_catalog.get_one()?,
                 c.command,
                 c.url,
                 c.engine,
@@ -363,10 +363,10 @@ pub fn get_command(
             Some(cli::SqlSubCommand::Server(sc)) => {
                 if sc.livy {
                     Box::new(SqlServerLivyCommand::new(
-                        work_catalog.get_one()?,
-                        work_catalog.get_one()?,
-                        work_catalog.get_one()?,
-                        work_catalog.get_one()?,
+                        cli_catalog.get_one()?,
+                        cli_catalog.get_one()?,
+                        cli_catalog.get_one()?,
+                        cli_catalog.get_one()?,
                         sc.address,
                         sc.port,
                     ))
@@ -376,7 +376,7 @@ pub fn get_command(
                             Box::new(SqlServerFlightSqlCommand::new(
                                 sc.address,
                                 sc.port,
-                                work_catalog.get_one()?,
+                                cli_catalog.get_one()?,
                             ))
                         } else {
                             return Err(CLIError::usage_error("Kamu was compiled without Flight SQL support"))
@@ -384,10 +384,10 @@ pub fn get_command(
                     }
                 } else {
                     Box::new(SqlServerCommand::new(
-                        work_catalog.get_one()?,
-                        work_catalog.get_one()?,
-                        work_catalog.get_one()?,
-                        work_catalog.get_one()?,
+                        cli_catalog.get_one()?,
+                        cli_catalog.get_one()?,
+                        cli_catalog.get_one()?,
+                        cli_catalog.get_one()?,
                         sc.address,
                         sc.port,
                     ))
@@ -397,20 +397,20 @@ pub fn get_command(
         cli::Command::System(c) => match c.subcommand {
             cli::SystemSubCommand::ApiServer(sc) => match sc.subcommand {
                 None => {
-                    let workspace_svc = work_catalog.get_one::<WorkspaceService>()?;
+                    let workspace_svc = cli_catalog.get_one::<WorkspaceService>()?;
 
                     Box::new(APIServerRunCommand::new(
                         base_catalog.clone(),
-                        work_catalog.clone(),
+                        cli_catalog.clone(),
                         workspace_svc.is_multi_tenant_workspace(),
-                        work_catalog.get_one()?,
+                        cli_catalog.get_one()?,
                         sc.address,
                         sc.http_port,
                         sc.external_address,
                         sc.get_token,
-                        work_catalog.get_one()?,
-                        work_catalog.get_one()?,
-                        work_catalog.get_one()?,
+                        cli_catalog.get_one()?,
+                        cli_catalog.get_one()?,
+                        cli_catalog.get_one()?,
                         args.e2e_output_data_path,
                     ))
                 }
@@ -422,11 +422,11 @@ pub fn get_command(
                 }
             },
             cli::SystemSubCommand::Compact(sc) => Box::new(CompactCommand::new(
-                work_catalog.get_one()?,
-                work_catalog.get_one()?,
-                work_catalog.get_one()?,
-                work_catalog.get_one()?,
-                validate_many_dataset_patterns(work_catalog, sc.dataset)?,
+                cli_catalog.get_one()?,
+                cli_catalog.get_one()?,
+                cli_catalog.get_one()?,
+                cli_catalog.get_one()?,
+                validate_many_dataset_patterns(cli_catalog, sc.dataset)?,
                 sc.max_slice_size,
                 sc.max_slice_records,
                 sc.hard,
@@ -434,53 +434,53 @@ pub fn get_command(
                 sc.keep_metadata_only,
             )),
             cli::SystemSubCommand::Diagnose(_) => Box::new(SystemDiagnoseCommand::new(
-                work_catalog.get_one()?,
-                work_catalog.get_one()?,
-                work_catalog.get_one()?,
-                work_catalog.get_one()?,
+                cli_catalog.get_one()?,
+                cli_catalog.get_one()?,
+                cli_catalog.get_one()?,
+                cli_catalog.get_one()?,
             )),
             cli::SystemSubCommand::DebugToken(sc) => {
-                Box::new(DebugTokenCommand::new(work_catalog.get_one()?, sc.token))
+                Box::new(DebugTokenCommand::new(cli_catalog.get_one()?, sc.token))
             }
             cli::SystemSubCommand::E2e(sc) => Box::new(SystemE2ECommand::new(
                 sc.action,
                 sc.dataset,
-                work_catalog.get_one()?,
+                cli_catalog.get_one()?,
             )),
-            cli::SystemSubCommand::Gc(_) => Box::new(GcCommand::new(work_catalog.get_one()?)),
+            cli::SystemSubCommand::Gc(_) => Box::new(GcCommand::new(cli_catalog.get_one()?)),
             cli::SystemSubCommand::GenerateToken(sc) => Box::new(GenerateTokenCommand::new(
-                work_catalog.get_one()?,
+                cli_catalog.get_one()?,
                 sc.login,
                 sc.subject,
                 sc.expiration_time_sec,
             )),
             cli::SystemSubCommand::Info(sc) => Box::new(SystemInfoCommand::new(
-                work_catalog.get_one()?,
-                work_catalog.get_one()?,
-                work_catalog.get_one()?,
+                cli_catalog.get_one()?,
+                cli_catalog.get_one()?,
+                cli_catalog.get_one()?,
                 sc.output_format,
             )),
             cli::SystemSubCommand::Ipfs(sc) => match sc.subcommand {
                 cli::SystemIpfsSubCommand::Add(ssc) => Box::new(SystemIpfsAddCommand::new(
-                    work_catalog.get_one()?,
+                    cli_catalog.get_one()?,
                     ssc.dataset,
                 )),
             },
             cli::SystemSubCommand::UpgradeWorkspace(_) => {
-                Box::new(UpgradeWorkspaceCommand::new(work_catalog.get_one()?))
+                Box::new(UpgradeWorkspaceCommand::new(cli_catalog.get_one()?))
             }
         },
         cli::Command::Tail(c) => Box::new(TailCommand::new(
-            work_catalog.get_one()?,
-            validate_dataset_ref(work_catalog, c.dataset)?,
+            cli_catalog.get_one()?,
+            validate_dataset_ref(cli_catalog, c.dataset)?,
             c.skip_records,
             c.num_records,
-            work_catalog.get_one()?,
+            cli_catalog.get_one()?,
         )),
         cli::Command::Ui(c) => {
-            let workspace_svc = work_catalog.get_one::<WorkspaceService>()?;
+            let workspace_svc = cli_catalog.get_one::<WorkspaceService>()?;
 
-            let current_account_subject = work_catalog.get_one::<CurrentAccountSubject>()?;
+            let current_account_subject = cli_catalog.get_one::<CurrentAccountSubject>()?;
 
             let current_account_name = match current_account_subject.as_ref() {
                 CurrentAccountSubject::Logged(l) => l.account_name.clone(),
@@ -493,30 +493,29 @@ pub fn get_command(
                 base_catalog.clone(),
                 workspace_svc.is_multi_tenant_workspace(),
                 current_account_name,
-                work_catalog.get_one()?,
-                work_catalog.get_one()?,
-                work_catalog.get_one()?,
-                work_catalog.get_one()?,
+                cli_catalog.get_one()?,
+                cli_catalog.get_one()?,
+                cli_catalog.get_one()?,
+                cli_catalog.get_one()?,
                 c.address,
                 c.http_port,
                 c.get_token,
             ))
         }
         cli::Command::Verify(c) => Box::new(VerifyCommand::new(
-            work_catalog.get_one()?,
-            work_catalog.get_one()?,
-            work_catalog.get_one()?,
-            work_catalog.get_one()?,
-            work_catalog.get_one()?,
-            validate_many_dataset_patterns(work_catalog, c.dataset)?.into_iter(),
+            cli_catalog.get_one()?,
+            cli_catalog.get_one()?,
+            cli_catalog.get_one()?,
+            cli_catalog.get_one()?,
+            cli_catalog.get_one()?,
+            validate_many_dataset_patterns(cli_catalog, c.dataset)?.into_iter(),
             c.recursive,
             c.integrity,
         )),
 
-        cli::Command::Version(c) => Box::new(VersionCommand::new(
-            work_catalog.get_one()?,
-            c.output_format,
-        )),
+        cli::Command::Version(c) => {
+            Box::new(VersionCommand::new(cli_catalog.get_one()?, c.output_format))
+        }
     };
 
     Ok(command)