+++ src/e2e/app/cli/inmem/tests/tests/mod.rs
@@ -12,4 +12,3 @@ mod test_auth;
 mod test_rest_api;
 mod test_selftest;
 mod test_smart_transfer_protocol;
-mod test_system_generate_token;