+++ CHANGELOG.md
@@ -18,6 +18,8 @@ Recommendation: for ease of reading, use the following order:
   - Add confirmation step to `system compact` command
   - Add support for patterns to `system compact` to process multiple datasets at once
   - Fixed argument parsing error in `kamu system compact` command
+### Fixed
+ - Broken catalog issue for server and transactional modes  
 
 ## [0.204.3] - 2024-09-26
 ### Fixed